Safari-611.1.10/LayoutTests/editing/text-iterator/basic-iteration-shadowdom-expected.txt

Unit tests for WebCore text iterator with shadow tree support enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addShadowTreeWithDivElement(div); range.selectNodeContents(internals.oldestShadowRoot(div)); internals.rangeAsText(range) is "b"
PASS appendBrElement(internals.oldestShadowRoot(div).childNodes[0]); range.selectNodeContents(internals.oldestShadowRoot(div)); internals.rangeAsText(range) is "b\n"
PASS successfullyParsed is true

TEST COMPLETE
